test: add update-tsconfig script
