+++ tsconfig.test.json
@@ -1,4 +1,8 @@
 {
   "extends": "./tsconfig.json",
-  "include": ["lib/**/*.ts", "test/**/*.ts"]
+  "include": [
+    "lib/**/*.ts",
+    "test/**/*.ts"
+  ],
+  "exclude": []
 }
